# openapi.yaml
openapi: 3.0.0

info:
  title: AccessMap Nairobi API
  version: 1.0.0
  description: >
    A community-powered accessibility directory for Nairobi venues.
    Exposes venue and report data for downstream partners such as Kanairo-Klean (Team 13).
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://accessmap-nairobi.onrender.com
    description: Production server (Render)

paths:

  /api/venues:
    get:
      summary: Get a list of venues, optionally filtered by category, area, and accessibility
      parameters:
        - name: category
          in: query
          required: false
          description: Filter venues by category e.g. market, hospital, school
          schema:
            type: string
            example: market
        - name: area
          in: query
          required: false
          description: Filter venues by Nairobi sub-area e.g. Westlands, Karen, CBD
          schema:
            type: string
            example: Westlands
        - name: accessible
          in: query
          required: false
          description: If true, only return venues with key accessibility features present
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: A list of venues matching the filters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Venue'
              example:
                - id: v_001
                  slug: kenyatta-national-hospital
                  name: Kenyatta National Hospital
                  area: Upper Hill
                  address: Hospital Rd, Nairobi
                  category: hospital
                  accessibility:
                    ramp: true
                    lift: true
                    accessibleToilet: true
                    accessibleParking: true
                    tactilePaving: false
                    wideCorridors: true
                    audioAssistance: false
                    staffAssistance: true
                  coordinates:
                    lat: -1.3004
                    lng: 36.8073
                  addedAt: "2024-01-15T10:30:00Z"
        '400':
          description: Invalid query parameter value supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Invalid query parameter value supplied

    post:
      summary: Add a new venue to the directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVenue'
            example:
              name: Zucchini Green Grocers
              area: Westlands
              address: Mpaka Rd, Westlands, Nairobi
              category: market
              accessibility:
                ramp: true
                lift: false
                accessibleToilet: false
                accessibleParking: true
                tactilePaving: false
                wideCorridors: true
                audioAssistance: false
                staffAssistance: false
              coordinates:
                lat: -1.2676
                lng: 36.8062
      responses:
        '201':
          description: Venue successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
              example:
                id: v_034
                slug: zucchini-green-grocers
                name: Zucchini Green Grocers
                area: Westlands
                address: Mpaka Rd, Westlands, Nairobi
                category: market
                accessibility:
                  ramp: true
                  lift: false
                  accessibleToilet: false
                  accessibleParking: true
                  tactilePaving: false
                  wideCorridors: true
                  audioAssistance: false
                  staffAssistance: false
                coordinates:
                  lat: -1.2676
                  lng: 36.8062
                addedAt: "2026-07-10T09:00:00Z"
        '400':
          description: Missing or invalid fields in request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: name is required and must be a string
        '409':
          description: A venue with this name already exists in this area
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                  - existing
                properties:
                  error:
                    type: string
                    example: A venue with this name already exists in this area.
                  existing:
                    $ref: '#/components/schemas/Venue'

  /api/venues/{slug}:
    get:
      summary: Get full details of a single venue by its slug
      parameters:
        - name: slug
          in: path
          required: true
          description: The URL-friendly slug of the venue e.g. kenyatta-national-hospital
          schema:
            type: string
            example: kenyatta-national-hospital
      responses:
        '200':
          description: Full venue details including address, coordinates and accessibility badges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
              example:
                id: v_001
                slug: kenyatta-national-hospital
                name: Kenyatta National Hospital
                area: Upper Hill
                address: Hospital Rd, Nairobi
                category: hospital
                accessibility:
                  ramp: true
                  lift: true
                  accessibleToilet: true
                  accessibleParking: true
                  tactilePaving: false
                  wideCorridors: true
                  audioAssistance: false
                  staffAssistance: true
                coordinates:
                  lat: -1.3004
                  lng: 36.8073
                addedAt: "2024-01-15T10:30:00Z"
        '404':
          description: No venue found with the given slug
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: No venue found with the given slug

  /api/venues/{slug}/rating:
    get:
      summary: Get the average community star rating for a specific venue
      parameters:
        - name: slug
          in: path
          required: true
          description: The URL-friendly slug of the venue
          schema:
            type: string
            example: kenyatta-national-hospital
      responses:
        '200':
          description: The average star rating for the venue
          content:
            application/json:
              schema:
                type: object
                required:
                  - slug
                  - averageRating
                  - totalReports
                properties:
                  slug:
                    type: string
                    example: kenyatta-national-hospital
                  averageRating:
                    type: number
                    format: float
                    description: Average rating rounded to one decimal place
                    example: 4.2
                  totalReports:
                    type: integer
                    description: Total number of community reports this rating is based on
                    example: 5
        '404':
          description: No venue found with the given slug
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: No venue found with the given slug

  /api/reports:
    post:
      summary: Submit a new community accessibility report for a venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewReport'
            example:
              venueId: v_001
              rating: 4
              description: The ramp at the entrance is well maintained and the staff were very helpful.
              visitedAt: "2026-06-15"
              accessibility:
                ramp: true
                lift: true
                accessibleToilet: true
                accessibleParking: true
                tactilePaving: false
                wideCorridors: true
                audioAssistance: false
                staffAssistance: true
      responses:
        '201':
          description: Report successfully submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
              example:
                id: r_021
                venueId: v_001
                rating: 4
                description: The ramp at the entrance is well maintained and the staff were very helpful.
                visitedAt: "2026-06-15"
                submittedAt: "2026-07-10T09:30:00Z"
                accessibility:
                  ramp: true
                  lift: true
                  accessibleToilet: true
                  accessibleParking: true
                  tactilePaving: false
                  wideCorridors: true
                  audioAssistance: false
                  staffAssistance: true
        '400':
          description: Missing or invalid fields in request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: rating is required and must be an integer from 1 to 5
        '404':
          description: No venue found with the given venueId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: No venue found with the given venueId

components:
  schemas:

    Error:
      type: object
      description: A validation or lookup error
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable explanation of what went wrong
          example: name is required and must be a string

    Accessibility:
      type: object
      description: The 8 accessibility features that can be present or absent at a venue
      required:
        - ramp
        - lift
        - accessibleToilet
        - accessibleParking
        - tactilePaving
        - wideCorridors
        - audioAssistance
        - staffAssistance
      properties:
        ramp:
          type: boolean
          description: Ramp or step-free entry available
          example: true
        lift:
          type: boolean
          description: Lift or elevator available
          example: false
        accessibleToilet:
          type: boolean
          description: Accessible toilet available
          example: true
        accessibleParking:
          type: boolean
          description: Accessible parking available
          example: true
        tactilePaving:
          type: boolean
          description: Tactile paving present for visually impaired visitors
          example: false
        wideCorridors:
          type: boolean
          description: Wide corridors that can accommodate wheelchairs
          example: true
        audioAssistance:
          type: boolean
          description: Audio assistance available
          example: false
        staffAssistance:
          type: boolean
          description: Staff trained to assist people with disabilities
          example: true

    Coordinates:
      type: object
      description: GPS coordinates of the venue
      required:
        - lat
        - lng
      properties:
        lat:
          type: number
          format: float
          description: Latitude
          example: -1.3004
        lng:
          type: number
          format: float
          description: Longitude
          example: 36.8073

    Venue:
      type: object
      description: A venue in the AccessMap Nairobi directory
      required:
        - id
        - slug
        - name
        - area
        - address
        - category
        - accessibility
        - addedAt
      properties:
        id:
          type: string
          description: Unique identifier for the venue
          example: v_001
        slug:
          type: string
          description: URL-friendly version of the venue name used in API paths
          example: kenyatta-national-hospital
        name:
          type: string
          description: Full name of the venue
          example: Kenyatta National Hospital
        area:
          type: string
          description: Nairobi sub-area where the venue is located
          example: Upper Hill
        address:
          type: string
          description: Street address of the venue
          example: Hospital Rd, Nairobi
        category:
          type: string
          description: Type of venue
          enum:
            - hospital
            - market
            - office
            - school
            - transport
            - bank
            - other
          example: hospital
        accessibility:
          $ref: '#/components/schemas/Accessibility'
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        addedAt:
          type: string
          format: date-time
          description: When the venue was added to the directory
          example: "2024-01-15T10:30:00Z"

    NewVenue:
      type: object
      description: The data required to create a new venue
      required:
        - name
        - area
        - address
        - category
        - accessibility
      properties:
        name:
          type: string
          example: Zucchini Green Grocers
        area:
          type: string
          example: Westlands
        address:
          type: string
          example: Mpaka Rd, Westlands, Nairobi
        category:
          type: string
          enum:
            - hospital
            - market
            - office
            - school
            - transport
            - bank
            - other
          example: market
        accessibility:
          $ref: '#/components/schemas/Accessibility'
        coordinates:
          $ref: '#/components/schemas/Coordinates'

    Report:
      type: object
      description: A community accessibility report submitted for a venue
      required:
        - id
        - venueId
        - rating
        - visitedAt
        - submittedAt
        - accessibility
      properties:
        id:
          type: string
          description: Unique identifier for the report
          example: r_021
        venueId:
          type: string
          description: The id of the venue this report belongs to
          example: v_001
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: Star rating from 1 to 5
          example: 4
        description:
          type: string
          description: Optional written description of the visitor's experience
          example: The ramp at the entrance is well maintained and staff were very helpful.
        visitedAt:
          type: string
          format: date
          description: The date the visitor visited the venue
          example: "2026-06-15"
        submittedAt:
          type: string
          format: date-time
          description: When the report was submitted to the system
          example: "2026-07-10T09:30:00Z"
        accessibility:
          $ref: '#/components/schemas/Accessibility'

    NewReport:
      type: object
      description: The data required to submit a new accessibility report
      required:
        - venueId
        - rating
        - visitedAt
        - accessibility
      properties:
        venueId:
          type: string
          description: >
            The id of the venue being reported on. Must be an id previously returned by GET /api/venues or GET /api/venues/{slug} — not invented by the caller.
          example: v_001
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: Star rating from 1 to 5
          example: 4
        description:
          type: string
          description: Optional written description of the experience
          example: The ramp at the entrance is well maintained and staff were very helpful.
        visitedAt:
          type: string
          format: date
          description: The date the visitor visited the venue
          example: "2026-06-15"
        accessibility:
          $ref: '#/components/schemas/Accessibility'